+++ lesson03-quyle/src/AppCar/UsingCarFunsion.java
@@ -2,17 +2,17 @@ package AppCar;
 
 public class UsingCarFunsion {
 	public static float calculate(UsingCar usingCar) {
-		float totalOfMoney = 0;
+		int totalOfMoney = 0;
 		Car us = usingCar.getCar();
 		totalOfMoney = us.getCost();
 		if(us.getCc()>200) {
-			totalOfMoney=(float) (totalOfMoney*0.05);
+			totalOfMoney=(int) (totalOfMoney*0.05);
 		}
 		if(us.getCc()<=200 && us.getCc()>=100) {
-			totalOfMoney=(float) (totalOfMoney*0.03);
+			totalOfMoney=(int) (totalOfMoney*0.03);
 		}
 		if(us.getCc()<100) {
-			totalOfMoney=(float) (totalOfMoney*0.01);
+			totalOfMoney=(int) (totalOfMoney*0.01);
 		}
 		return totalOfMoney;
 	}
